backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.11.1
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.3.1
dnspython==2.7.0
exceptiongroup==1.2.2
Flask==3.1.3
Flask-Cors==6.0.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
motor==3.7.0
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.11.0
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
Werkzeug==3.1.7
a2wsgi==1.10.10